requirements.txt
pandas
numpy
sqlalchemy
pyodbc
python-dotenv
jupyterlab
notebook
matplotlib
seaborn
plotly
openpyxl
